Gutted Vector3. Added extension methods; JIT
JIT and C# language specs cause a lot of issues with floating point.
Unity3D isn't right, I'm not right and Wolfram isn't right. There is
going to be error... Sooo the API has changed slightly.

A lot of the API is now extension methods. That should let JIT do its'
magic. It also lets me define, and not define, certain functionality for
Types of vectors.