Rewrote ClassBuilder and added Builder Detail Providers
This is some ugly code and there were only so many ways it could be
written. Rosyln produces some intricate and cool code but it's not the
prettiest and setup of runtime generation of source code requires a lot
of hoops to be jumped through.

This is hopefully a concise and OOP enough version that will support all
our use-cases.